Major optimizations
Any bpy.ops calls automatically update the scene, which leads to a
massive amount of overhead. Most of the calls are now replaced
with direct copies. Unfortunately, adding hooks was the most expensive
part and there does not seem to be a way to accomplish the hooking
without the switch to edit mode, which has become the limiting step.
We should maybe consider adding an option to remove hooking so the
rest of everything loads quickly.